cloud: Move join related functions to their own file
The cloud.go file is getting a bit too long, and future patches would
have added even more to it.  This patch pulls out all of the code
needed by the join() method into its own file.
